requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
psycopg2-binary
gunicorn==21.2.0
python-dotenv==1.0.1
boto3==1.34.0
Pillow